closes #23 refactored addpicture form
